Fix build failure with Zephyr SDK
This will fix some of the build failures after upgrading to Zephyr
1.14.0 LTS and the new 0.10.0 SDK.

+++ src/zjs_pwm.c
@@ -108,8 +108,8 @@ static ZJS_DECL_FUNC(zjs_pwm_pin_set_cycles)
     }
 
     // update the JS object
-    double period = periodHW / sys_clock_hw_cycles_per_sec * 1000;
-    double pulseWidth = pulseWidthHW / sys_clock_hw_cycles_per_sec * 1000;
+    double period = periodHW / (int)sys_clock_hw_cycles_per_sec * 1000;
+    double pulseWidth = pulseWidthHW / (int)sys_clock_hw_cycles_per_sec * 1000;
     zjs_obj_add_number(this, "period", period);
     zjs_obj_add_number(this, "pulseWidth", pulseWidth);
